Support `--out-link` in `nix flake check`
Related to: #13470

+++ tests/functional/flakes/check.sh
@@ -158,8 +158,14 @@ expectStderr 0 nix flake check "$flakeDir" | grepQuiet 'running 1 flake check'
 
 # Test that `nix flake check --print-out-paths` produces the same out path as `nix build --print-out-paths`
 outPath=$(nix flake check "$flakeDir" --print-out-paths)
-outPathBuild=$(nix build "${flakeDir}#checks.$system.foo" --print-out-paths)
+outPathBuild=$(nix build "${flakeDir}#checks.$system.foo" --print-out-paths --no-link)
 [[ "$outPath" = "$outPathBuild" ]] || fail "out paths from flake check and build don't match"
+
+# Test out links
+! test -e result || fail "unexpected out link result found"
+nix flake check "$flakeDir" --out-link result
+test -e result || fail "out link result not found"
+rm result
 
 cat > "$flakeDir"/flake.nix <<EOF
 {
